rearrangements of header file, removal of Allocator
