+++ operators/cvcuda_holoscan_interop/cvcuda_to_holoscan.cpp
@@ -21,6 +21,7 @@
 
 #include "cvcuda_to_holoscan.hpp"
 #include "cvcuda_utils.hpp"
+#include "gxf_utils.hpp"
 #include "holoscan/holoscan.hpp"
 
 namespace holoscan::ops {
@@ -28,7 +29,6 @@ namespace holoscan::ops {
 void CvCudaToHoloscan::setup(OperatorSpec& spec) {
   spec.input<nvcv::Tensor>("input");
   spec.output<holoscan::TensorMap>("output");
-  spec.param(pool_, "pool", "Pool", "Pool to allocate the Holoscan tensor");
 }
 
 void CvCudaToHoloscan::compute(InputContext& op_input, OutputContext& op_output,
@@ -42,26 +42,8 @@ void CvCudaToHoloscan::compute(InputContext& op_input, OutputContext& op_output,
 
   HOLOSCAN_LOG_DEBUG("before create_out_message_with_tensor_like");
   const auto& [out_message, tensor_data_pointer] =
-      create_out_message_with_tensor_like(context.context(), cv_in_tensor, pool_);
+      create_out_message_with_tensor(context.context(), cv_in_tensor);
   HOLOSCAN_LOG_DEBUG("create_out_message_with_tensor_like success");
-
-  //   nvcv::TensorDataStridedCuda::Buffer cv_out_buffer;
-  //   auto in_strided_data = cv_in_tensor.exportData<nvcv::TensorDataStridedCuda>();
-  //   cv_out_buffer.basePtr = static_cast<NVCVByte*>(*tensor_data_pointer);
-  //   for (int i = 0; i < cv_in_tensor.rank(); i++) {
-  //     cv_out_buffer.strides[i] = in_strided_data->stride(i);
-  //   }
-
-  //   nvcv::TensorDataStridedCuda out_data(cv_in_tensor.shape(), cv_in_tensor.dtype(),
-  //   cv_out_buffer);
-
-  //   nvcv::Tensor cv_out_tensor = nvcv::TensorWrapData(out_data);
-
-  //   std::cout << "cv_out_tensor shape: " << cv_out_tensor.shape() << std::endl;
-
-  //   cvcuda::Reformat reformatOp;
-  //   reformatOp(0, cv_in_tensor, cv_out_tensor);
-  //   HOLOSCAN_LOG_DEBUG("ReformatOp success");
 
   op_output.emit(out_message, "output");
 }
